requirements.txt
mysql-connector-python==8.0.23
mysqlclient==2.0.3
protobuf==3.14.0
PyQt5==5.15.2
PyQt5-sip==12.8.1
six==1.15.0
SQLAlchemy==1.3.23
